+++ deployment/src/main/java/com/tietoevry/quarkus/resteasy/problem/deployment/RestEasyClassicDetector.java
@@ -2,12 +2,12 @@ package com.tietoevry.quarkus.resteasy.problem.deployment;
 
 import java.util.function.BooleanSupplier;
 
-final class RESTeasyReactiveDetector implements BooleanSupplier {
+final class RestEasyClassicDetector implements BooleanSupplier {
 
     @Override
     public boolean getAsBoolean() {
         try {
-            Class.forName("io.quarkus.resteasy.reactive.server.runtime.QuarkusContextProducers");
+            Class.forName("io.quarkus.resteasy.common.runtime.ResteasyContextProvider");
             return true;
         } catch (ClassNotFoundException e) {
             return false;
